ERC20 Prices page now a RSC and its form, a client
